Adds plan transmutation to archived plans, creates inframappingname key
